requirements.txt
langchain==0.0.177
gpt4all==0.2.3
chromadb==0.3.23
pdfminer.six==20221105
python-dotenv==1.0.0


